Enum optional::OptionBool

pub enum OptionBool {
    SomeTrue,
    SomeFalse,
    None,
}

The OptionBool type, a space-efficient Option replacement

Variants

SomeTrue

Some(true)

SomeFalse

Some(false)

None

None

Implementations

impl OptionBool

pub fn some(b: bool) -> Self

Create a SomeTrue for true, SomeFalse for false

pub fn none() -> Self

Create a None value.

Examples

assert_eq!(OptionBool::none(), optional::OptionBool::None);

pub fn is_some(&self) -> bool

Returns true if the option is a Some value.

Examples

assert!(OptionBool::SomeTrue.is_some());
assert!(OptionBool::SomeFalse.is_some());
assert!(!OptionBool::None.is_some());

pub fn is_none(&self) -> bool

Returns true if the option is a Some value.

Examples

assert!(!OptionBool::SomeTrue.is_none());
assert!(!OptionBool::SomeFalse.is_none());
assert!(OptionBool::None.is_none());

pub fn expect(&self, msg: &str) -> bool

Unwraps the contained bool, panics on None with given message.

Panics

if self is None

Examples

For SomeTrue/SomeFalse, the corresponding bool is returned.

assert!(OptionBool::SomeTrue.expect("FAIL"));
assert!(!OptionBool::SomeFalse.expect("FAIL"));

On None, it panics with the given message.

OptionBool::None.expect("FAIL"); // panics with FAIL

pub fn unwrap(&self) -> bool

Unwraps the contained bool, panics on None.

Panics

if self is None

Examples

For SomeTrue/SomeFalse, the corresponding bool is returned.

assert!(OptionBool::SomeTrue.unwrap());
assert!(!OptionBool::SomeFalse.unwrap());

On None, it panics with “unwrap called on None”

OptionBool::None.unwrap(); // panics

pub fn unwrap_or(&self, def: bool) -> bool

Returns the contained bool or a default.

Examples

assert!(OptionBool::SomeTrue.unwrap_or(false));
assert!(!OptionBool::SomeFalse.unwrap_or(true));
assert!(OptionBool::None.unwrap_or(true));
assert!(!OptionBool::None.unwrap_or(false));

pub fn unwrap_or_else<F>(self, f: F) -> boolwhere F: FnOnce() -> bool,

Returns the contained bool or a computed default.

Examples

assert!(OptionBool::SomeTrue.unwrap_or_else(|| false));
assert!(!OptionBool::SomeFalse.unwrap_or_else(|| panic!()));
assert!(OptionBool::None.unwrap_or_else(|| true));

pub fn map<U, F>(self, f: F) -> Option<U>where F: FnOnce(bool) -> U,

Maps an OptionBool to an Option<U> by applying the function over the contained bool.

Note that there is also map_bool(..) which works similarly, but returns another OptionBool.

Examples

Convert the contained bool to a Yes/No message

assert_eq!(Some("Yes"), OptionBool::SomeTrue.map(
    |b| if b { "Yes" } else { "No" }));

pub fn map_bool<F>(self, f: F) -> OptionBoolwhere F: FnOnce(bool) -> bool,

Maps an OptionBool to another OptionBool by applying the function over the contained bool.

Note that there is also map(..) which works similarly, but returns an Option<bool>.

Examples

Invert the contained bool

assert_eq!(OptionBool::SomeTrue,
    OptionBool::SomeFalse.map_bool(|b| !b));

pub fn map_or<U, F>(self, default: U, f: F) -> Uwhere F: FnOnce(bool) -> U,

Maps a value to a U by applying the function or return a default U.

Examples

Map to a string (as per the daily wtf’s boolean definition):

assert_eq!("True", OptionBool::SomeTrue.map_or("FileNotFound",
    |b| if b { "True" } else { "False" }));

pub fn map_or_else<U, D, F>(self, default: D, f: F) -> Uwhere D: FnOnce() -> U, F: FnOnce(bool) -> U,

Maps a value to a U by applying the function or return a computed default.

Examples

assert_eq!("True", OptionBool::SomeTrue.map_or_else(|| "FileNotFound",
    |b| if b { "True" } else { "False" }));

pub fn ok_or<E>(self, err: E) -> Result<bool, E>

Transforms the OptionBool into a Result<bool, E>, mapping SomeX to Ok(X) and None to Err(err).

Examples

assert_eq!(OptionBool::SomeTrue.ok_or("Ouch"), Ok(true));
assert_eq!(OptionBool::None.ok_or("Ouch"), Err("Ouch"));

pub fn ok_or_else<E, F>(self, err: F) -> Result<bool, E>where F: FnOnce() -> E,

Transforms the OptionBool into a Result<bool, E>, mapping SomeX to Ok(X) and None to a calculated Err(err).

Examples

assert_eq!(OptionBool::SomeTrue.ok_or_else(|| something_expensive()), Ok(true));
assert_eq!(OptionBool::None.ok_or_else(|| "Ouch"), Err("Ouch"));

pub fn and<U>(self, optb: Option<U>) -> Option<U>

Returns None if the option is None, otherwise returns optb.

Examples

assert_eq!(Some(1), OptionBool::SomeTrue.and(Some(1)));
assert_eq!(None, OptionBool::None.and(Some(1)));
let actual : Option<u8> = None;
assert_eq!(None, OptionBool::SomeTrue.and(actual));

pub fn and_bool(self, optb: OptionBool) -> OptionBool

Returns None if the option is None, otherwise returns optb.

Examples

assert_eq!(OptionBool::SomeTrue,
    OptionBool::SomeFalse.and_bool(OptionBool::SomeTrue));
assert_eq!(OptionBool::None,
    OptionBool::None.and_bool(OptionBool::SomeFalse));
assert_eq!(OptionBool::None,
    OptionBool::SomeTrue.and_bool(OptionBool::None));

pub fn and_then<U, F>(self, f: F) -> Option<U>where F: FnOnce(bool) -> Option<U>,

returns None if the OptionBool is None, otherwise calls f with the boolean value and returns the result as an Option<U>

Note that there is also and_then_bool(..) which works similarly, but returns another OptionBool.

Examples

assert_eq!(None, OptionBool::SomeFalse.and_then(
    |x| if x { Some(true) } else { None }));

pub fn and_then_bool<F>(self, f: F) -> OptionBoolwhere F: FnOnce(bool) -> OptionBool,

returns None if the OptionBool is None, otherwise calls f with the boolean value and returns the result as an OptionBool

Note that there is also and_then(..) which works similarly, but returns an Option<bool>.

Examples

assert_eq!(OptionBool::None, OptionBool::SomeFalse.and_then_bool(
    |x| if x { OptionBool::SomeTrue } else { OptionBool::None }));

pub fn or(self, optb: Option<bool>) -> Option<bool>

Returns this as Option unless this is None, in which case returns optb.

Examples

assert_eq!(Some(false), OptionBool::SomeFalse.or(Some(true)));
assert_eq!(Some(true), OptionBool::None.or(Some(true)));
assert_eq!(None, OptionBool::None.or(None));

pub fn or_bool(self, optb: OptionBool) -> OptionBool

Returns this as Option unless this is None, in which case returns optb.

Examples

assert_eq!(OptionBool::SomeFalse,
    OptionBool::SomeFalse.or_bool(OptionBool::SomeTrue));
assert_eq!(OptionBool::SomeTrue,
    OptionBool::None.or_bool(OptionBool::SomeTrue));
assert_eq!(OptionBool::None,
    OptionBool::None.or_bool(OptionBool::None));

pub fn or_else<F>(self, f: F) -> Option<bool>where F: FnOnce() -> Option<bool>,

Returns this as Option unless this is None, in which case use the supplied function to calculate the result.

Note that there is also or_else_bool(..) which works similarly, but returns another OptionBool.

Examples

assert_eq!(Some(false), OptionBool::SomeFalse.or_else(|| Some(true)));
assert_eq!(Some(true), OptionBool::None.or_else(|| Some(true)));
assert_eq!(None, OptionBool::None.or_else(|| None));

pub fn or_else_bool<F>(self, f: F) -> OptionBoolwhere F: FnOnce() -> OptionBool,

Returns this as Option unless this is None, in which case use the supplied function to calculate the result.

Note that there is also or_else(..) which works similarly, but returns an Option<bool>.

Examples

assert_eq!(OptionBool::SomeFalse,
    OptionBool::SomeFalse.or_else_bool(|| OptionBool::SomeTrue));
assert_eq!(OptionBool::SomeTrue,
    OptionBool::None.or_else_bool(|| OptionBool::SomeTrue));
assert_eq!(OptionBool::None,
    OptionBool::None.or_else_bool(|| OptionBool::None));

pub fn iter(&self) -> Iter<'_, bool>

return an iterator over all contained (that is zero or one) values.

Examples

assert_eq!(None, OptionBool::None.iter().next());
assert_eq!(Some(&true), OptionBool::SomeTrue.iter().next());

pub fn as_slice(self) -> &'static [bool]

return a possibly empty slice with the contained value, if any.

Examples

assert_eq!(&[true], OptionBool::SomeTrue.as_slice());
assert!(OptionBool::None.as_slice().is_empty());

pub fn take(&mut self) -> Option<bool>

Takes the value out of the OptionBool and returns ist as Option<bool>, changing self to None.

Note that there is also take_bool(..) which works similarly, but returns an OptionBool.

Examples

let mut x = OptionBool::some(true);
assert_eq!(Some(true), x.take());
assert_eq!(OptionBool::None, x);

pub fn take_bool(&mut self) -> OptionBool

Takes the value out of the OptionBool, changing self to None.

Note that there is also take(..) which works similarly, but returns an Option<bool>.

Examples

let mut x = OptionBool::some(true);
assert_eq!(OptionBool::some(true), x.take_bool());
assert_eq!(OptionBool::None, x);