Trait optional::Noned

pub trait Noned {
    fn is_none(&self) -> bool;
    fn get_none() -> Self;
}

A trait whose implementation for any type T allows the use of Optioned<T> where T is bound by both Sized and Copy.

Required Methods

fn is_none(&self) -> bool

Returns true if the contained value is the declared None for T, false otherwise.

fn get_none() -> Self

Returns the declared None value for T.